Remove dynamic classification
