+++ sdk/textanalytics/Azure.AI.TextAnalytics/src/TextAnalyticsClient.cs
@@ -2196,194 +2196,6 @@ namespace Azure.AI.TextAnalytics
 
         #endregion
 
-        #region Dynamic Classify
-
-        /// <summary>
-        /// Performs dynamic classification on the given document.
-        /// For document length limits, maximum batch size, and supported text encoding, see
-        /// <see href="https://aka.ms/azsdk/textanalytics/data-limits"/>.
-        /// </summary>
-        /// <remarks>
-        /// This method is only available for <see cref="TextAnalyticsClientOptions.ServiceVersion.V2022_10_01_Preview"/>, and newer.
-        /// </remarks>
-        /// <param name="document">The document to analyze.</param>
-        /// <param name="categories">The categories that the documents can be classified with.</param>
-        /// <param name="classificationType">The type of classification to perform.</param>
-        /// <param name="language">The language that the documents are written in.</param>
-        /// <param name="cancellationToken">The <see cref="CancellationToken"/> controlling the lifetime of the request.</param>
-        /// <returns>
-        /// The collection of categories used to classify each document that was successfully analyzed.
-        /// </returns>
-        /// <exception cref="NotSupportedException">This method is only supported in service API version 2022-10-01-preview and newer.</exception>
-        /// <exception cref="RequestFailedException">The service returned a non-success status code.</exception>
-        /// <exception cref="ArgumentException"><paramref name="document"/> is an empty string or <paramref name="categories"/> is an empty collection.</exception>
-        /// <exception cref="ArgumentNullException"><paramref name="document"/> or <paramref name="categories"/> is null.</exception>
-        public virtual Response<ClassificationCategoryCollection> DynamicClassify(
-            string document,
-            IEnumerable<string> categories,
-            ClassificationType? classificationType = default,
-            string language = default,
-            CancellationToken cancellationToken = default) =>
-            _serviceClient.DynamicClassify(document, categories, classificationType, language, cancellationToken);
-
-        /// <summary>
-        /// Performs dynamic classification on the given documents.
-        /// For document length limits, maximum batch size, and supported text encoding, see
-        /// <see href="https://aka.ms/azsdk/textanalytics/data-limits"/>.
-        /// </summary>
-        /// <remarks>
-        /// This method is only available for <see cref="TextAnalyticsClientOptions.ServiceVersion.V2022_10_01_Preview"/>, and newer.
-        /// </remarks>
-        /// <param name="documents">The documents to analyze.</param>
-        /// <param name="categories">The categories that the documents can be classified with.</param>
-        /// <param name="classificationType">The type of classification to perform.</param>
-        /// <param name="language">The language that the documents are written in.</param>
-        /// <param name="options">The additional <see cref="TextAnalyticsRequestOptions"/> used to configure the operation.</param>
-        /// <param name="cancellationToken">The <see cref="CancellationToken"/> controlling the lifetime of the request.</param>
-        /// <returns>
-        /// The collection of categories used to classify each document that was successfully analyzed.
-        /// </returns>
-        /// <exception cref="NotSupportedException">This method is only supported in service API version 2022-10-01-preview and newer.</exception>
-        /// <exception cref="RequestFailedException">The service returned a non-success status code.</exception>
-        /// <exception cref="ArgumentException"><paramref name="documents"/> or <paramref name="categories"/> is an empty collection.</exception>
-        /// <exception cref="ArgumentNullException"><paramref name="documents"/> or <paramref name="categories"/> is null.</exception>
-        public virtual Response<DynamicClassifyDocumentResultCollection> DynamicClassifyBatch(
-            IEnumerable<string> documents,
-            IEnumerable<string> categories,
-            ClassificationType? classificationType = default,
-            string language = default,
-            TextAnalyticsRequestOptions options = default,
-            CancellationToken cancellationToken = default)
-        {
-            options?.CheckSupported(ServiceVersion);
-            return _serviceClient.DynamicClassifyBatch(documents, categories, classificationType, language, options, cancellationToken);
-        }
-
-        /// <summary>
-        /// Performs dynamic classification on the given documents.
-        /// For document length limits, maximum batch size, and supported text encoding, see
-        /// <see href="https://aka.ms/azsdk/textanalytics/data-limits"/>.
-        /// </summary>
-        /// <remarks>
-        /// This method is only available for <see cref="TextAnalyticsClientOptions.ServiceVersion.V2022_10_01_Preview"/>, and newer.
-        /// </remarks>
-        /// <param name="documents">The documents to analyze.</param>
-        /// <param name="categories">The categories that the documents can be classified with.</param>
-        /// <param name="classificationType">The type of classification to perform.</param>
-        /// <param name="options">The additional <see cref="TextAnalyticsRequestOptions"/> used to configure the operation.</param>
-        /// <param name="cancellationToken">The <see cref="CancellationToken"/> controlling the lifetime of the request.</param>
-        /// <returns>
-        /// The collection of categories used to classify each document that was successfully analyzed.
-        /// </returns>
-        /// <exception cref="NotSupportedException">This method is only supported in service API version 2022-10-01-preview and newer.</exception>
-        /// <exception cref="RequestFailedException">The service returned a non-success status code.</exception>
-        /// <exception cref="ArgumentException"><paramref name="documents"/> or <paramref name="categories"/> is an empty collection.</exception>
-        /// <exception cref="ArgumentNullException"><paramref name="documents"/> or <paramref name="categories"/> is null.</exception>
-        public virtual Response<DynamicClassifyDocumentResultCollection> DynamicClassifyBatch(
-            IEnumerable<TextDocumentInput> documents,
-            IEnumerable<string> categories,
-            ClassificationType? classificationType = default,
-            TextAnalyticsRequestOptions options = default,
-            CancellationToken cancellationToken = default)
-        {
-            options?.CheckSupported(ServiceVersion);
-            return _serviceClient.DynamicClassifyBatch(documents, categories, classificationType, options, cancellationToken);
-        }
-
-        /// <summary>
-        /// Performs dynamic classification on the given document.
-        /// For document length limits, maximum batch size, and supported text encoding, see
-        /// <see href="https://aka.ms/azsdk/textanalytics/data-limits"/>.
-        /// </summary>
-        /// <remarks>
-        /// This method is only available for <see cref="TextAnalyticsClientOptions.ServiceVersion.V2022_10_01_Preview"/>, and newer.
-        /// </remarks>
-        /// <param name="document">The document to analyze.</param>
-        /// <param name="categories">The categories that the documents can be classified with.</param>
-        /// <param name="classificationType">The type of classification to perform.</param>
-        /// <param name="language">The language that the documents are written in.</param>
-        /// <param name="cancellationToken">The <see cref="CancellationToken"/> controlling the lifetime of the request.</param>
-        /// <returns>
-        /// The collection of categories used to classify each document that was successfully analyzed.
-        /// </returns>
-        /// <exception cref="NotSupportedException">This method is only supported in service API version 2022-10-01-preview and newer.</exception>
-        /// <exception cref="RequestFailedException">The service returned a non-success status code.</exception>
-        /// <exception cref="ArgumentException"><paramref name="document"/> is an empty string or <paramref name="categories"/> is an empty collection.</exception>
-        /// <exception cref="ArgumentNullException"><paramref name="document"/> or <paramref name="categories"/> is null.</exception>
-        public virtual async Task<Response<ClassificationCategoryCollection>> DynamicClassifyAsync(
-            string document,
-            IEnumerable<string> categories,
-            ClassificationType? classificationType = default,
-            string language = default,
-            CancellationToken cancellationToken = default) =>
-            await _serviceClient.DynamicClassifyAsync(document, categories, classificationType, language, cancellationToken).ConfigureAwait(false);
-
-        /// <summary>
-        /// Performs dynamic classification on the given documents.
-        /// For document length limits, maximum batch size, and supported text encoding, see
-        /// <see href="https://aka.ms/azsdk/textanalytics/data-limits"/>.
-        /// </summary>
-        /// <remarks>
-        /// This method is only available for <see cref="TextAnalyticsClientOptions.ServiceVersion.V2022_10_01_Preview"/>, and newer.
-        /// </remarks>
-        /// <param name="documents">The documents to analyze.</param>
-        /// <param name="categories">The categories that the documents can be classified with.</param>
-        /// <param name="classificationType">The type of classification to perform.</param>
-        /// <param name="language">The language that the documents are written in.</param>
-        /// <param name="options">The additional <see cref="TextAnalyticsRequestOptions"/> used to configure the operation.</param>
-        /// <param name="cancellationToken">The <see cref="CancellationToken"/> controlling the lifetime of the request.</param>
-        /// <returns>
-        /// The collection of categories used to classify each document that was successfully analyzed.
-        /// </returns>
-        /// <exception cref="NotSupportedException">This method is only supported in service API version 2022-10-01-preview and newer.</exception>
-        /// <exception cref="RequestFailedException">The service returned a non-success status code.</exception>
-        /// <exception cref="ArgumentException"><paramref name="documents"/> or <paramref name="categories"/> is an empty collection.</exception>
-        /// <exception cref="ArgumentNullException"><paramref name="documents"/> or <paramref name="categories"/> is null.</exception>
-        public virtual async Task<Response<DynamicClassifyDocumentResultCollection>> DynamicClassifyBatchAsync(
-            IEnumerable<string> documents,
-            IEnumerable<string> categories,
-            ClassificationType? classificationType = default,
-            string language = default,
-            TextAnalyticsRequestOptions options = default,
-            CancellationToken cancellationToken = default)
-        {
-            options?.CheckSupported(ServiceVersion);
-            return await _serviceClient.DynamicClassifyBatchAsync(documents, categories, classificationType, language, options,cancellationToken).ConfigureAwait(false);
-        }
-
-        /// <summary>
-        /// Performs dynamic classification on the given documents.
-        /// For document length limits, maximum batch size, and supported text encoding, see
-        /// <see href="https://aka.ms/azsdk/textanalytics/data-limits"/>.
-        /// </summary>
-        /// <remarks>
-        /// This method is only available for <see cref="TextAnalyticsClientOptions.ServiceVersion.V2022_10_01_Preview"/>, and newer.
-        /// </remarks>
-        /// <param name="documents">The documents to analyze.</param>
-        /// <param name="categories">The categories that the documents can be classified with.</param>
-        /// <param name="classificationType">The type of classification to perform.</param>
-        /// <param name="options">The additional <see cref="TextAnalyticsRequestOptions"/> used to configure the operation.</param>
-        /// <param name="cancellationToken">The <see cref="CancellationToken"/> controlling the lifetime of the request.</param>
-        /// <returns>
-        /// The collection of categories used to classify each document that was successfully analyzed.
-        /// </returns>
-        /// <exception cref="NotSupportedException">This method is only supported in service API version 2022-10-01-preview and newer.</exception>
-        /// <exception cref="RequestFailedException">The service returned a non-success status code.</exception>
-        /// <exception cref="ArgumentException"><paramref name="documents"/> or <paramref name="categories"/> is an empty collection.</exception>
-        /// <exception cref="ArgumentNullException"><paramref name="documents"/> or <paramref name="categories"/> is null.</exception>
-        public virtual async Task<Response<DynamicClassifyDocumentResultCollection>> DynamicClassifyBatchAsync(
-            IEnumerable<TextDocumentInput> documents,
-            IEnumerable<string> categories,
-            ClassificationType? classificationType = default,
-            TextAnalyticsRequestOptions options = default,
-            CancellationToken cancellationToken = default)
-        {
-            options?.CheckSupported(ServiceVersion);
-            return await _serviceClient.DynamicClassifyBatchAsync(documents, categories, classificationType, options, cancellationToken).ConfigureAwait(false);
-        }
-
-        #endregion
-
         #region Extract Summary
 
         /// <summary>
