+++ sdk/textanalytics/Azure.AI.TextAnalytics/src/Transforms.cs
@@ -497,41 +497,6 @@ namespace Azure.AI.TextAnalytics
 
         #endregion
 
-        #region Dynamic Classify
-
-        internal static ClassificationCategoryCollection ConvertToClassificationCategoryCollection(DynamicClassificationResultDocumentsItem document)
-        {
-            return new ClassificationCategoryCollection(ConvertToClassificationCategoryList(document.Classifications.ToList()), ConvertToWarnings(document.Warnings));
-        }
-
-        internal static DynamicClassifyDocumentResultCollection ConvertToDynamicClassifyDocumentResultCollection(DynamicClassificationResult result, IDictionary<string, int> idToIndexMap)
-        {
-            var documentResults = new List<ClassifyDocumentResult>(result.Documents.Count);
-
-            // Read errors.
-            foreach (InputError error in result.Errors)
-            {
-                documentResults.Add(new ClassifyDocumentResult(error.Id, ConvertToError(error.Error)));
-            }
-
-            // Read results.
-            foreach (DynamicClassificationResultDocumentsItem document in result.Documents)
-            {
-                documentResults.Add(
-                    new ClassifyDocumentResult(
-                        document.Id,
-                        document.Statistics ?? default,
-                        ConvertToClassificationCategoryCollection(document),
-                        detectedLanguage: default,
-                        ConvertToWarnings(document.Warnings)));
-            }
-
-            documentResults = SortHeterogeneousCollection(documentResults, idToIndexMap);
-            return new DynamicClassifyDocumentResultCollection(documentResults, result.Statistics, result.ModelVersion);
-        }
-
-        #endregion
-
         #region Extract Summary
 
         internal static List<SummarySentence> ConvertToSummarySentenceList(IEnumerable<ExtractedSummarySentence> sentences)
