--- sdk/textanalytics/Azure.AI.TextAnalytics/tests/SessionRecords/DynamicClassifyTests/DynamicClassifyBatchConvenienceTestAsync.json
@@ -1,114 +1,0 @@
-{
-  "Entries": [
-    {
-      "RequestUri": "https://javatextanalyticstestresources.cognitiveservices.azure.com/language/:analyze-text?api-version=2022-10-01-preview\u0026showStats=false",
-      "RequestMethod": "POST",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Content-Length": "292",
-        "Content-Type": "application/json",
-        "Ocp-Apim-Subscription-Key": "Sanitized",
-        "traceparent": "00-353d6b6cf4cbeba829ebccff5e3dcf75-60f167f3b3b3fdaa-00",
-        "User-Agent": "azsdk-net-AI.TextAnalytics/5.3.0-alpha.20221129.1 (.NET 6.0.11; Microsoft Windows 10.0.22621)",
-        "x-ms-client-request-id": "065a51f11014c36eb53ddc7feed05e52",
-        "x-ms-return-client-request-id": "true"
-      },
-      "RequestBody": {
-        "analysisInput": {
-          "documents": [
-            {
-              "id": "0",
-              "text": "The WHO is issuing a warning about monkeypox.",
-              "language": "en"
-            },
-            {
-              "id": "1",
-              "text": "Mo Salah plays in Liverpool FC in England.",
-              "language": "en"
-            }
-          ]
-        },
-        "parameters": {
-          "categories": [
-            "Health",
-            "Politics",
-            "Music",
-            "Sports"
-          ]
-        },
-        "kind": "DynamicClassification"
-      },
-      "StatusCode": 200,
-      "ResponseHeaders": {
-        "apim-request-id": "c3ebad3d-1501-4e46-aec0-8cfc34aac7ff",
-        "Content-Length": "561",
-        "Content-Type": "application/json; charset=utf-8",
-        "csp-billing-usage": "CognitiveServices.TextAnalytics.BatchScoring=0,CognitiveServices.TextAnalytics.TextRecords=2",
-        "Date": "Wed, 30 Nov 2022 00:38:31 GMT",
-        "Set-Cookie": ".AspNetCore.Mvc.CookieTempDataProvider=; expires=Thu, 01 Jan 1970 00:00:00 GMT; path=/; samesite=lax; httponly",
-        "Strict-Transport-Security": "max-age=31536000; includeSubDomains; preload",
-        "X-Content-Type-Options": "nosniff",
-        "x-envoy-upstream-service-time": "273",
-        "x-ms-region": "West US 2"
-      },
-      "ResponseBody": {
-        "kind": "DynamicClassificationResults",
-        "results": {
-          "documents": [
-            {
-              "id": "0",
-              "classifications": [
-                {
-                  "category": "Health",
-                  "confidenceScore": 0.8
-                },
-                {
-                  "category": "Sports",
-                  "confidenceScore": 0.07
-                },
-                {
-                  "category": "Music",
-                  "confidenceScore": 0.07
-                },
-                {
-                  "category": "Politics",
-                  "confidenceScore": 0.06
-                }
-              ],
-              "warnings": []
-            },
-            {
-              "id": "1",
-              "classifications": [
-                {
-                  "category": "Sports",
-                  "confidenceScore": 0.99
-                },
-                {
-                  "category": "Music",
-                  "confidenceScore": 0.0
-                },
-                {
-                  "category": "Health",
-                  "confidenceScore": 0.0
-                },
-                {
-                  "category": "Politics",
-                  "confidenceScore": 0.0
-                }
-              ],
-              "warnings": []
-            }
-          ],
-          "errors": [],
-          "modelVersion": "2022-10-01-preview"
-        }
-      }
-    }
-  ],
-  "Variables": {
-    "RandomSeed": "770907455",
-    "TEXT_ANALYTICS_API_KEY": "Sanitized",
-    "TEXT_ANALYTICS_ENDPOINT": "https://javatextanalyticstestresources.cognitiveservices.azure.com"
-  }
-}
