--- sdk/textanalytics/Azure.AI.TextAnalytics/tests/SessionRecords/DynamicClassifyTests/DynamicClassifyBatchConvenienceWithErrorTest.json
@@ -1,104 +1,0 @@
-{
-  "Entries": [
-    {
-      "RequestUri": "https://javatextanalyticstestresources.cognitiveservices.azure.com/language/:analyze-text?api-version=2022-10-01-preview\u0026showStats=false",
-      "RequestMethod": "POST",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Content-Length": "247",
-        "Content-Type": "application/json",
-        "Ocp-Apim-Subscription-Key": "Sanitized",
-        "traceparent": "00-8677c03abf7968cd41a584f651ed03f2-1037e8d0f3b7712c-00",
-        "User-Agent": "azsdk-net-AI.TextAnalytics/5.3.0-alpha.20221129.1 (.NET 6.0.11; Microsoft Windows 10.0.22621)",
-        "x-ms-client-request-id": "c9876054639918882b825113eb4669cf",
-        "x-ms-return-client-request-id": "true"
-      },
-      "RequestBody": {
-        "analysisInput": {
-          "documents": [
-            {
-              "id": "0",
-              "text": "Mo Salah plays in Liverpool FC in England.",
-              "language": "en"
-            },
-            {
-              "id": "1",
-              "text": "",
-              "language": "en"
-            }
-          ]
-        },
-        "parameters": {
-          "categories": [
-            "Health",
-            "Politics",
-            "Music",
-            "Sports"
-          ]
-        },
-        "kind": "DynamicClassification"
-      },
-      "StatusCode": 200,
-      "ResponseHeaders": {
-        "apim-request-id": "a190b67d-a52f-4788-8d0f-976cb0d37000",
-        "Content-Length": "498",
-        "Content-Type": "application/json; charset=utf-8",
-        "csp-billing-usage": "CognitiveServices.TextAnalytics.BatchScoring=0,CognitiveServices.TextAnalytics.TextRecords=1",
-        "Date": "Wed, 30 Nov 2022 00:38:06 GMT",
-        "Set-Cookie": ".AspNetCore.Mvc.CookieTempDataProvider=; expires=Thu, 01 Jan 1970 00:00:00 GMT; path=/; samesite=lax; httponly",
-        "Strict-Transport-Security": "max-age=31536000; includeSubDomains; preload",
-        "X-Content-Type-Options": "nosniff",
-        "x-envoy-upstream-service-time": "198",
-        "x-ms-region": "West US 2"
-      },
-      "ResponseBody": {
-        "kind": "DynamicClassificationResults",
-        "results": {
-          "documents": [
-            {
-              "id": "0",
-              "classifications": [
-                {
-                  "category": "Sports",
-                  "confidenceScore": 0.99
-                },
-                {
-                  "category": "Music",
-                  "confidenceScore": 0.0
-                },
-                {
-                  "category": "Health",
-                  "confidenceScore": 0.0
-                },
-                {
-                  "category": "Politics",
-                  "confidenceScore": 0.0
-                }
-              ],
-              "warnings": []
-            }
-          ],
-          "errors": [
-            {
-              "id": "1",
-              "error": {
-                "code": "InvalidArgument",
-                "message": "Invalid Document in request.",
-                "innererror": {
-                  "code": "InvalidDocument",
-                  "message": "Document text is empty."
-                }
-              }
-            }
-          ],
-          "modelVersion": "2022-10-01-preview"
-        }
-      }
-    }
-  ],
-  "Variables": {
-    "RandomSeed": "866505865",
-    "TEXT_ANALYTICS_API_KEY": "Sanitized",
-    "TEXT_ANALYTICS_ENDPOINT": "https://javatextanalyticstestresources.cognitiveservices.azure.com"
-  }
-}
