--- sdk/textanalytics/Azure.AI.TextAnalytics/tests/SessionRecords/DynamicClassifyTests/DynamicClassifyBatchWithDisableServiceLogsAsync.json
@@ -1,115 +1,0 @@
-{
-  "Entries": [
-    {
-      "RequestUri": "https://javatextanalyticstestresources.cognitiveservices.azure.com/language/:analyze-text?api-version=2022-10-01-preview\u0026showStats=false",
-      "RequestMethod": "POST",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Content-Length": "313",
-        "Content-Type": "application/json",
-        "Ocp-Apim-Subscription-Key": "Sanitized",
-        "traceparent": "00-4b963305d9a08f32b63c08bd9f29bd2a-318e9ff18e5451f4-00",
-        "User-Agent": "azsdk-net-AI.TextAnalytics/5.3.0-alpha.20221129.1 (.NET 6.0.11; Microsoft Windows 10.0.22621)",
-        "x-ms-client-request-id": "e72804136a4be1fad21c739f1c74c923",
-        "x-ms-return-client-request-id": "true"
-      },
-      "RequestBody": {
-        "analysisInput": {
-          "documents": [
-            {
-              "id": "0",
-              "text": "The WHO is issuing a warning about monkeypox.",
-              "language": "en"
-            },
-            {
-              "id": "1",
-              "text": "Mo Salah plays in Liverpool FC in England.",
-              "language": "en"
-            }
-          ]
-        },
-        "parameters": {
-          "categories": [
-            "Health",
-            "Politics",
-            "Music",
-            "Sports"
-          ],
-          "loggingOptOut": true
-        },
-        "kind": "DynamicClassification"
-      },
-      "StatusCode": 200,
-      "ResponseHeaders": {
-        "apim-request-id": "80a2de5a-3b3b-4b7b-96e5-47ff87f14350",
-        "Content-Length": "561",
-        "Content-Type": "application/json; charset=utf-8",
-        "csp-billing-usage": "CognitiveServices.TextAnalytics.BatchScoring=0,CognitiveServices.TextAnalytics.TextRecords=2",
-        "Date": "Wed, 30 Nov 2022 00:45:05 GMT",
-        "Set-Cookie": ".AspNetCore.Mvc.CookieTempDataProvider=; expires=Thu, 01 Jan 1970 00:00:00 GMT; path=/; samesite=lax; httponly",
-        "Strict-Transport-Security": "max-age=31536000; includeSubDomains; preload",
-        "X-Content-Type-Options": "nosniff",
-        "x-envoy-upstream-service-time": "389",
-        "x-ms-region": "West US 2"
-      },
-      "ResponseBody": {
-        "kind": "DynamicClassificationResults",
-        "results": {
-          "documents": [
-            {
-              "id": "0",
-              "classifications": [
-                {
-                  "category": "Health",
-                  "confidenceScore": 0.8
-                },
-                {
-                  "category": "Sports",
-                  "confidenceScore": 0.07
-                },
-                {
-                  "category": "Music",
-                  "confidenceScore": 0.07
-                },
-                {
-                  "category": "Politics",
-                  "confidenceScore": 0.06
-                }
-              ],
-              "warnings": []
-            },
-            {
-              "id": "1",
-              "classifications": [
-                {
-                  "category": "Sports",
-                  "confidenceScore": 0.99
-                },
-                {
-                  "category": "Music",
-                  "confidenceScore": 0.0
-                },
-                {
-                  "category": "Health",
-                  "confidenceScore": 0.0
-                },
-                {
-                  "category": "Politics",
-                  "confidenceScore": 0.0
-                }
-              ],
-              "warnings": []
-            }
-          ],
-          "errors": [],
-          "modelVersion": "2022-10-01-preview"
-        }
-      }
-    }
-  ],
-  "Variables": {
-    "RandomSeed": "759308561",
-    "TEXT_ANALYTICS_API_KEY": "Sanitized",
-    "TEXT_ANALYTICS_ENDPOINT": "https://javatextanalyticstestresources.cognitiveservices.azure.com"
-  }
-}
