--- sdk/textanalytics/Azure.AI.TextAnalytics/tests/SessionRecords/DynamicClassifyTests/DynamicClassifyBatchWithNullIdTestAsync.json
@@ -1,64 +1,0 @@
-{
-  "Entries": [
-    {
-      "RequestUri": "https://javatextanalyticstestresources.cognitiveservices.azure.com/language/:analyze-text?api-version=2022-10-01-preview\u0026showStats=false",
-      "RequestMethod": "POST",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Content-Length": "180",
-        "Content-Type": "application/json",
-        "Ocp-Apim-Subscription-Key": "Sanitized",
-        "traceparent": "00-31ec1fb2453ce9fe8b7b3e4e38056527-737292315dc9cfb6-00",
-        "User-Agent": "azsdk-net-AI.TextAnalytics/5.3.0-alpha.20221129.1 (.NET 6.0.11; Microsoft Windows 10.0.22621)",
-        "x-ms-client-request-id": "6d8aaa4f4e35cc3764c5c060282619df",
-        "x-ms-return-client-request-id": "true"
-      },
-      "RequestBody": {
-        "analysisInput": {
-          "documents": [
-            {
-              "id": null,
-              "text": "Hello world",
-              "language": "en"
-            }
-          ]
-        },
-        "parameters": {
-          "categories": [
-            "Health",
-            "Politics",
-            "Music",
-            "Sports"
-          ]
-        },
-        "kind": "DynamicClassification"
-      },
-      "StatusCode": 400,
-      "ResponseHeaders": {
-        "apim-request-id": "7b83ba9c-cea9-4147-9cd7-5524160d08a4",
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Wed, 30 Nov 2022 00:38:35 GMT",
-        "Strict-Transport-Security": "max-age=31536000; includeSubDomains; preload",
-        "Transfer-Encoding": "chunked",
-        "X-Content-Type-Options": "nosniff",
-        "x-envoy-upstream-service-time": "5",
-        "x-ms-region": "West US 2"
-      },
-      "ResponseBody": {
-        "error": {
-          "code": "InvalidRequest",
-          "message": "Invalid Document in request.",
-          "innererror": {
-            "code": "InvalidDocument",
-            "message": "At least one input is missing an Id attribute."
-          }
-        }
-      }
-    }
-  ],
-  "Variables": {
-    "RandomSeed": "1820872791",
-    "TEXT_ANALYTICS_API_KEY": "Sanitized",
-    "TEXT_ANALYTICS_ENDPOINT": "https://javatextanalyticstestresources.cognitiveservices.azure.com"
-  }
-}
