+++ sdk/textanalytics/Azure.AI.TextAnalytics/tests/samples/Sample11_ExtractSummaryConvenience.cs
