+++ sdk/textanalytics/Azure.AI.TextAnalytics/tests/samples/Sample12_AbstractSummaryConvenience.cs
